@@ -30,67 +30,68 @@
     "coverage": "vitest run --coverage"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "3.1.0",
-    "@eslint/js": "9.10.0",
-    "@playwright/test": "1.47.1",
-    "@quramy/prisma-fabbrica": "2.2.1",
-    "@sveltejs/adapter-auto": "3.2.4",
-    "@sveltejs/adapter-vercel": "5.4.3",
-    "@sveltejs/kit": "2.5.26",
-    "@sveltejs/vite-plugin-svelte": "3.1.2",
-    "@tailwindcss/forms": "0.5.9",
-    "@testing-library/jest-dom": "6.5.0",
+    "@eslint/eslintrc": "3.2.0",
+    "@eslint/js": "9.18.0",
+    "@playwright/test": "1.49.1",
+    "@quramy/prisma-fabbrica": "2.2.2",
+    "@sveltejs/adapter-auto": "4.0.0",
+    "@sveltejs/adapter-vercel": "5.5.3",
+    "@sveltejs/kit": "2.15.3",
+    "@sveltejs/vite-plugin-svelte": "5.0.3",
+    "@tailwindcss/forms": "0.5.10",
+    "@testing-library/jest-dom": "6.6.3",
     "@types/gtag.js": "0.0.20",
-    "@typescript-eslint/eslint-plugin": "8.5.0",
-    "@typescript-eslint/parser": "8.5.0",
-    "@vitest/coverage-v8": "2.0.5",
-    "@vitest/ui": "2.0.5",
-    "eslint": "9.10.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-svelte": "2.43.0",
-    "flowbite-svelte": "0.46.6",
-    "flowbite-svelte-icons": "1.6.1",
-    "globals": "15.9.0",
-    "husky": "9.1.6",
-    "jsdom": "25.0.0",
-    "lint-staged": "15.2.10",
-    "prettier": "3.3.3",
-    "prettier-plugin-svelte": "3.2.6",
-    "prettier-plugin-tailwindcss": "0.6.6",
-    "prisma": "5.19.1",
+    "@typescript-eslint/eslint-plugin": "8.20.0",
+    "@typescript-eslint/parser": "8.20.0",
+    "@vitest/coverage-v8": "3.0.0-beta.3",
+    "@vitest/ui": "3.0.0-beta.3",
+    "eslint": "9.18.0",
+    "eslint-config-prettier": "10.0.1",
+    "eslint-plugin-svelte": "2.46.1",
+    "flowbite-svelte": "0.47.4",
+    "flowbite-svelte-icons": "2.0.2",
+    "globals": "15.14.0",
+    "husky": "9.1.7",
+    "jsdom": "26.0.0",
+    "lint-staged": "15.3.0",
+    "nock": "13.5.6",
+    "prettier": "3.4.2",
+    "prettier-plugin-svelte": "3.3.3",
+    "prettier-plugin-tailwindcss": "0.6.10",
+    "prisma": "5.22.0",
     "stwui": "0.21.2-next",
-    "svelte": "4.2.19",
-    "svelte-check": "4.0.2",
-    "svelte-meta-tags": "3.1.4",
-    "sveltekit-superforms": "2.17.0",
-    "tslib": "2.7.0",
-    "typescript": "5.6.2",
-    "vite": "5.4.5",
-    "vitest": "2.0.5",
-    "zod": "3.23.8"
+    "super-sitemap": "1.0.3",
+    "svelte": "5.19.0",
+    "svelte-check": "4.1.4",
+    "svelte-meta-tags": "4.1.0",
+    "sveltekit-superforms": "2.22.1",
+    "tslib": "2.8.1",
+    "typescript": "5.6.3",
+    "vite": "6.0.7",
+    "vitest": "3.0.0-beta.3",
+    "zod": "3.24.1"
   },
   "type": "module",
   "dependencies": {
     "@lucia-auth/adapter-prisma": "3.0.2",
     "@mermaid-js/mermaid-cli": "10.9.1",
     "@popperjs/core": "2.11.8",
-    "@prisma/client": "5.19.1",
-    "@testing-library/svelte": "5.2.1",
-    "@types/jest": "29.5.13",
+    "@prisma/client": "5.22.0",
+    "@testing-library/svelte": "5.2.6",
+    "@types/jest": "29.5.14",
+    "@types/node": "22.10.7",
     "autoprefixer": "10.4.20",
-    "debug": "4.3.7",
-    "flowbite": "2.5.1",
-    "just-extend": "6.2.0",
+    "debug": "4.4.0",
+    "flowbite": "2.5.2",
     "lucia": "2.7.7",
-    "playwright": "1.47.1",
-    "pnpm": "9.10.0",
+    "playwright": "1.49.1",
+    "pnpm": "9.15.4",
     "prisma-erd-generator": "1.11.2",
-    "svelte-add": "2023.9.17-0.0",
-    "svelte-eslint-parser": "^0.41.1",
-    "tailwind-merge": "2.5.2",
-    "tailwindcss": "3.4.11",
-    "vercel": "37.4.2",
+    "svelte-eslint-parser": "0.43.0",
+    "tailwind-merge": "2.6.0",
+    "tailwindcss": "3.4.17",
+    "vercel": "39.3.0",
     "xss": "1.0.15"
   },
-  "packageManager": "pnpm@9.1.0"
+  "packageManager": "pnpm@9.15.2"
 }
