@@ -18,8 +18,8 @@
   import ExternalLinkWrapper from '$lib/components/ExternalLinkWrapper.svelte';
   import AcceptedCounter from '$lib/components/SubmissionStatus/AcceptedCounter.svelte';
   import { getBackgroundColorFrom } from '$lib/services/submission_status';
-  import { getContestNameLabel } from '$lib/utils/contest';
-  import { taskUrl, toWhiteTextIfNeeds } from '$lib/utils/task';
+  import { addContestNameToTaskIndex } from '$lib/utils/contest';
+  import { toWhiteTextIfNeeds, getTaskUrl, removeTaskIndexFromTitle } from '$lib/utils/task';
 
   export let grade: string;
   export let gradeColor: string;
@@ -53,45 +53,50 @@
     <!-- TODO: 問題が多くなってきたら、ページネーションを導入する -->
     <!-- TODO: 回答状況に応じて、フィルタリングできるようにする -->
     <div class="overflow-auto rounded-md border">
-      <Table shadow id={grade} class="text-md">
+      <Table shadow id={grade} class="text-md table-fixed w-full">
         <TableHead class="text-sm bg-gray-100">
-          <TableHeadCell class="min-w-[96px] max-w-[120px]">回答</TableHeadCell>
-          <TableHeadCell class="min-w-[240px] max-w-2/3 text-left pl-0 sm:pl-6 truncate">
+          <TableHeadCell class="w-20 sm:w-24 min-w-[5rem]">回答</TableHeadCell>
+          <TableHeadCell class="w-1/2 text-left pl-0 sm:pl-6 truncate overflow-ellipsis">
             問題名
           </TableHeadCell>
-          <TableHeadCell class="min-w-[120px] max-w-[150px] text-left pl-0 truncate">
+          <TableHeadCell class="w-1/3 hidden sm:table-cell text-left pl-0 truncate">
             出典
           </TableHeadCell>
-          <TableHeadCell class="min-w-[24px] text-center">
+          <TableHeadCell class="w-6 text-center">
             <span class="sr-only">編集</span>
           </TableHeadCell>
         </TableHead>
         <TableBody tableBodyClass="divide-y">
           {#each taskResults as taskResult}
             <TableBodyRow
-              key={taskResult.contest_id + '-' + taskResult.task_id}
+              id={taskResult.contest_id + '-' + taskResult.task_id}
               class={getBackgroundColorFrom(taskResult.status_name)}
             >
               <TableBodyCell
-                class="justify-center w-20 px-0.5 sm:px-3"
-                on:click={() => updatingModal.openModal(taskResult)}
+                class="justify-center w-20 px-1 sm:px-3 pt-1 sm:pt-3 pb-0.5 sm:pb-1"
+                onclick={() => updatingModal.openModal(taskResult)}
               >
-                <div class="flex items-center justify-center min-w-[80px] max-w-[80px]">
+                <div class="flex items-center justify-center w-full h-full">
                   <SubmissionStatusImage {taskResult} {isLoggedIn} />
                 </div>
               </TableBodyCell>
-              <TableBodyCell class="pl-0 sm:pl-6 w-2/3">
+              <TableBodyCell class="w-1/2 text-left truncate pl-0 sm:pl-6">
                 <ExternalLinkWrapper
-                  url={taskUrl(taskResult.contest_id, taskResult.task_id)}
-                  description={taskResult.title}
+                  url={getTaskUrl(taskResult.contest_id, taskResult.task_id)}
+                  description={removeTaskIndexFromTitle(
+                    taskResult.title,
+                    taskResult.task_table_index,
+                  )}
                   textSize="xs:text-lg"
                   textColorInDarkMode="dark:text-gray-300"
                 />
               </TableBodyCell>
-              <TableBodyCell class="pl-0 xs:text-lg text-gray-700 dark:text-gray-300">
-                {getContestNameLabel(taskResult.contest_id)}
+              <TableBodyCell
+                class="w-1/3 hidden sm:table-cell text-left truncate pl-0 xs:text-lg text-gray-700 dark:text-gray-300"
+              >
+                {addContestNameToTaskIndex(taskResult.contest_id, taskResult.task_table_index)}
               </TableBodyCell>
-              <TableBodyCell class="px-0">
+              <TableBodyCell class="w-6 px-0">
                 {#if isAdmin}
                   <div class="flex justify-center items-center px-0">
                     <a
