@@ -17,10 +17,11 @@
   }
 </script>
 
-<Img src={imagePath} alt={imageAlt} class="h-8 w-8" />
+<Img src={imagePath} alt={imageAlt} class="h-7 xs:h-8 w-7 xs:w-8" />
+
 {#if isLoggedIn}
   <div class="flex flex-col items-center ml-2 md:ml-4 text-xs">
-    <div class="mb-1 dark:text-gray-300">
+    <div class="pb-1 dark:text-gray-300">
       {'更新'}
     </div>
     <ChevronDownOutline class="w-3 h-3 text-primary-600 dark:text-gray-300 inline" />
