@@ -4,10 +4,24 @@ export const ATCODER_BASE_CONTEST_URL: string = `${ATCODER_BASE_URL}/contests`;
 
 export const ATCODER_PROBLEMS_URL: string = 'https://kenkoooo.com/atcoder/#/';
 
+export const ATCODER_PROBLEMS_API_BASE_URL: string = 'https://kenkoooo.com/atcoder/resources/';
+
+export const AOJ_API_BASE_URL: string = 'https://judgeapi.u-aizu.ac.jp/';
+
+// Note:
+// AIZU ONLINE JUDGE (AOJ) has multiple versions: v1.0, v2.0, and v3.0.
+// As of late October 2024, we are using v2.0, which appears to be the main site.
+export const AOJ_BASE_URL: string = 'https://onlinejudge.u-aizu.ac.jp';
+
+export const AOJ_TASKS_URL: string = `${AOJ_BASE_URL}/problems`;
+
 export const AOJ_ATCODER_JOI_URL: string = 'https://joi.goodbaton.com/';
 
 export const AOJ_PCK_URL: string = 'https://pro-ktmr.github.io/aoj-pck/';
 
-export const AOJ_ICPC_URL: string = 'http://aoj-icpc.ichyo.jp/';
+export const ICPC_JAPAN_PROBLEMS_URL: string = 'https://icpc-japan-problems.irrrrr.cc/';
 
 export const GITHUB_URL: string = 'https://github.com/AtCoder-NoviSteps/AtCoderNoviSteps';
+
+export const TASK_GRADE_CRITERIA_SHEET_URL: string =
+  'https://docs.google.com/spreadsheets/d/1GJbTRRBsoBaY-CXIr3dIXmxkwacV4nHOTOIMCmo__Ug/';
