@@ -1,6 +1,6 @@
-import '@testing-library/jest-dom';
-import { getTasks } from './services/problemsApiService';
 import { test } from 'vitest';
+
+import { getTasks } from '$lib/clients';
 
 test('call problems api', () => {
   getTasks();
