@@ -1,72 +1,128 @@
 import { redirect, type Actions } from '@sveltejs/kit';
 
-import { type Contest, type Task, type ImportTask, TaskGrade, getTaskGrade } from '$lib/types/task';
+import { Roles } from '$lib/types/user';
+import type { Contests, ContestForImport, ContestsForImport } from '$lib/types/contest';
+import {
+  type Task,
+  type Tasks,
+  type TaskForImport,
+  type TasksForImport,
+  TaskGrade,
+  getTaskGrade,
+} from '$lib/types/task';
+
 import * as taskService from '$lib/services/tasks';
 import * as userService from '$lib/services/users';
-import * as problemApiService from '$lib/services/problemsApiService';
+import * as apiClient from '$lib/clients';
 
+import { isAdmin } from '$lib/utils/authorship';
 import { sha256 } from '$lib/utils/hash';
 import { classifyContest } from '$lib/utils/contest';
 
-import { Roles } from '$lib/types/user';
+import { LOGIN_PAGE } from '$lib/constants/navbar-links';
+import { TEMPORARY_REDIRECT } from '$lib/constants/http-response-status-codes.js';
 
 export async function load({ locals }) {
+  await validateAdminAccess(locals);
+
+  const { contestsForImport, tasksForImport } = await fetchContestsAndTasksFromAPI();
+
+  const tasksFromDB = await taskService.getTasks();
+  const registeredTaskMap = prepareTaskMap(tasksFromDB);
+
+  const unregisteredTasks = filterUnregisteredTasks(
+    contestsForImport,
+    tasksForImport,
+    registeredTaskMap,
+  );
+  const contestsWithUnregisteredTasks: Contests = mergeContestsAndUnregisteredTasks(
+    contestsForImport,
+    unregisteredTasks,
+  );
+
+  return {
+    importContests: contestsWithUnregisteredTasks,
+  };
+}
+
+async function validateAdminAccess(locals: App.Locals): Promise<void> {
   const session = await locals.auth.validate();
+
   if (!session) {
-    redirect(302, '/login');
+    redirect(TEMPORARY_REDIRECT, LOGIN_PAGE);
   }
 
   const user = await userService.getUser(session?.user.username as string);
-  if (user?.role !== Roles.ADMIN) {
-    redirect(302, '/login');
+
+  if (!isAdmin(user?.role as Roles)) {
+    redirect(TEMPORARY_REDIRECT, LOGIN_PAGE);
   }
+}
 
-  const importContestsJson = await problemApiService.getContests();
-  const importTasksJson = await problemApiService.getTasks();
-  const tasks = await taskService.getTasks();
+async function fetchContestsAndTasksFromAPI(): Promise<{
+  contestsForImport: ContestsForImport;
+  tasksForImport: TasksForImport;
+}> {
+  const contestsForImport = await apiClient.getContests();
+  const tasksForImport = await apiClient.getTasks();
 
-  //dbから取得した、contest_id-Task, task_id-Taskのマップ
-  const taskContestMap = new Map<string, Task>();
+  return { contestsForImport, tasksForImport };
+}
+
+function prepareTaskMap(tasks: Tasks): Map<string, Task> {
   const taskMap = new Map<string, Task>();
-  for (let i = 0; i < tasks.length; i++) {
-    taskContestMap.set(tasks[i].contest_id, tasks[i]);
-    taskMap.set(tasks[i].task_id, tasks[i]);
-  }
-  //APIから取得した、contest_id-ImportTaskのマップ
-  const unregisteredTasksInContest = new Map<string, ImportTask[]>();
 
-  //対象コンテストに絞る
-  // See: src/lib/utils/contest.ts
-  for (let i = 0; i < importContestsJson.length; i++) {
-    const contest_id = importContestsJson[i].id;
-    const contest_type = classifyContest(contest_id);
-
-    if (contest_type === null) {
-      continue;
-    }
-
-    unregisteredTasksInContest.set(
-      contest_id,
-      importTasksJson.filter(
-        (importTaskJson: ImportTask) =>
-          importTaskJson.contest_id == contest_id && !taskMap.has(importTaskJson.id),
-      ),
-    );
-  }
-
-  const importContests = importContestsJson.map((importContestJson: Contest) => {
-    return {
-      id: importContestJson.id,
-      title: importContestJson.title,
-      start_epoch_second: importContestJson.start_epoch_second,
-      duration_second: importContestJson.duration_second,
-      tasks: unregisteredTasksInContest.get(importContestJson.id) ?? [],
-    };
+  tasks.forEach((task: Task) => {
+    taskMap.set(task.task_id, task);
   });
 
-  return {
-    importContests: importContests,
-  };
+  return taskMap;
+}
+
+// See:
+// src/lib/utils/contest.ts
+function filterUnregisteredTasks(
+  contestsForImport: ContestsForImport,
+  tasksForImport: TasksForImport,
+  registeredTaskMap: Map<string, Task>,
+): Map<string, TasksForImport> {
+  const unregisteredTasks = new Map<string, TasksForImport>();
+
+  contestsForImport.forEach((contestForImport: ContestForImport) => {
+    const contest_id = contestForImport.id;
+    const contest_type = classifyContest(contest_id);
+
+    if (contest_type !== null) {
+      unregisteredTasks.set(
+        contest_id,
+        tasksForImport.filter(
+          (taskForImport: TaskForImport) =>
+            taskForImport.contest_id == contest_id && !registeredTaskMap.has(taskForImport.id),
+        ),
+      );
+    }
+  });
+
+  return unregisteredTasks;
+}
+
+function mergeContestsAndUnregisteredTasks(
+  contestsForImport: ContestsForImport,
+  unregisteredTasks: Map<string, TasksForImport>,
+) {
+  const contestsWithUnregisteredTasks: Contests = contestsForImport.map(
+    (contestForImport: ContestForImport) => {
+      return {
+        id: contestForImport.id,
+        title: contestForImport.title,
+        start_epoch_second: contestForImport.start_epoch_second,
+        duration_second: contestForImport.duration_second,
+        tasks: unregisteredTasks.get(contestForImport.id) ?? [],
+      };
+    },
+  );
+
+  return contestsWithUnregisteredTasks;
 }
 
 export const actions: Actions = {
@@ -76,10 +132,12 @@ export const actions: Actions = {
       const formData = await request.formData();
       const contest_id = formData.get('contest_id')?.toString() as string;
 
-      const tasks = await problemApiService.getTasks();
-      const tasksByContestId = tasks.filter((task: ImportTask) => task.contest_id === contest_id);
+      const tasks = await apiClient.getTasks();
+      const tasksByContestId = tasks.filter(
+        (task: TaskForImport) => task.contest_id === contest_id,
+      );
 
-      tasksByContestId.map(async (task: ImportTask) => {
+      tasksByContestId.map(async (task: TaskForImport) => {
         const id = (await sha256(contest_id + task.title)) as string;
         await taskService.createTask(id, task.id, task.contest_id, task.problem_index, task.title);
       });
@@ -124,11 +182,13 @@ export const actions: Actions = {
       await taskService.updateTask(task_id, task_grade);
       const contest_id = formData.get('contest_id')?.toString() as string;
 
-      const tasks = await problemApiService.getTasks();
+      const tasks = await apiClient.getTasks();
 
-      const tasksByContestId = tasks.filter((task: ImportTask) => task.contest_id === contest_id);
+      const tasksByContestId = tasks.filter(
+        (task: TaskForImport) => task.contest_id === contest_id,
+      );
 
-      tasksByContestId.map(async (task: ImportTask) => {
+      tasksByContestId.map(async (task: TaskForImport) => {
         const id = (await sha256(contest_id + task.title)) as string;
         console.log(id);
         await taskService.createTask(id, task.id, task.contest_id, task.problem_index, task.title);
