@@ -2,7 +2,7 @@
   import HeadingOne from '$lib/components/HeadingOne.svelte';
   import TaskListForEdit from '$lib/components/TaskListForEdit.svelte';
   export let data;
-  //let tasks = data.tasks;
+
   let importContests = data.importContests;
 </script>
 
