@@ -1,22 +1,25 @@
 <script lang="ts">
   // See:
   // https://github.com/oekazuma/svelte-meta-tags
-  import { page } from '$app/stores';
-  import { MetaTags } from 'svelte-meta-tags';
-  import extend from 'just-extend';
+  // https://oekazuma.github.io/svelte-meta-tags/ja/migration-guide/
+  import { page } from '$app/state';
+  import { navigating } from '$app/stores';
+
+  import { MetaTags, deepMerge } from 'svelte-meta-tags';
 
   import '../app.css';
 
   import Header from '$lib/components/Header.svelte';
   import GoogleAnalytics from '$lib/components/GoogleAnalytics.svelte';
   import ErrorMessageToast from '$lib/components/ToastWrapper/ErrorMessageToast.svelte';
+  import SpinnerWrapper from '$lib/components/SpinnerWrapper.svelte';
   import Footer from '$lib/components/Footer.svelte';
 
   import { errorMessageStore } from '$lib/stores/error_message';
 
   export let data;
 
-  $: metaTags = extend(true, {}, data.baseMetaTags, $page.data.pageMetaTags);
+  $: metaTags = deepMerge(data.baseMetaTags, page.data.pageMetaTags);
 </script>
 
 <Header />
@@ -26,6 +29,12 @@
 
 <ErrorMessageToast errorMessage={$errorMessageStore} />
 
-<slot />
+<!-- See: -->
+<!-- https://svelte.dev/docs/kit/$app-stores#navigating -->
+{#if $navigating}
+  <SpinnerWrapper />
+{:else}
+  <slot />
+{/if}
 
 <Footer />
